Migration from SVN complete
This status shows many comments in stantard output and has a delay on the start with big logs.
Not ready for distribution.
